requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-deepseek
langserve
pydantic
python-dotenv
sse_starlette